Split bitmap_font from code_08003980. Split bitmap_font_data from data_08936b54. Manage includes accordingly.
