+++ data/bitmap_font_data.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "global.h"
+#include "src/bitmap_font.h"
+#include "src/task_pool.h"
+
+extern const char *D_08936b50; // Fullwidth Lowercase Latin Alphabet Table
+extern struct TaskMethods D_08936b54;
+extern struct TaskMethods D_08936b64;
+extern struct TaskMethods D_08936b74;
