+++ src/code_0800b778.c
@@ -1,7 +1,7 @@
 #include "global.h"
 #include "code_0800b778.h"
 #include "code_08001360.h"
-#include "code_08003980.h"
+#include "bitmap_font.h"
 #include "memory_heap.h"
 #include "code_08007468.h"
 #include "src/lib_08049144.h"
