+++ src/engines/fireworks.c
@@ -1,7 +1,7 @@
 #include "engines/fireworks.h"
 
 #include "src/code_08001360.h"
-#include "src/code_08003980.h"
+#include "src/bitmap_font.h"
 #include "src/code_08007468.h"
 #include "src/text_printer.h"
 #include "src/code_0800b778.h"
