+++ src/scenes/game_select.c
@@ -2,7 +2,7 @@
 #include "game_select.h"
 #include "src/memory.h"
 #include "src/code_08001360.h"
-#include "src/code_08003980.h"
+#include "src/bitmap_font.h"
 #include "src/memory_heap.h"
 #include "src/code_08007468.h"
 #include "src/text_printer.h"
