+++ src/scenes/main_menu.c
@@ -3,7 +3,7 @@
 
 #include "src/main.h"
 #include "src/code_08001360.h"
-#include "src/code_08003980.h"
+#include "src/bitmap_font.h"
 #include "src/memory_heap.h"
 #include "src/code_080068f8.h"
 #include "src/code_08007468.h"
