+++ src/scenes/results.c
@@ -4,7 +4,7 @@
 #include "src/main.h"
 #include "src/memory.h"
 #include "src/code_08001360.h"
-#include "src/code_08003980.h"
+#include "src/bitmap_font.h"
 #include "src/memory_heap.h"
 #include "src/code_08007468.h"
 #include "src/text_printer.h"
